activemq-rpc/pythonStompTest/requirements.txt
docopt==0.6.2
pkg-resources==0.0.0
stomp.py==4.1.19
